+++ src/binary/audio/synth/stop_ch.c
@@ -1,7 +1,7 @@
 /**
  * Stop Channel Playback
  */
-void init_ch_state(int *ch_state) {
+void init_ch_state(long *ch_state) {
   *(ch_state + 0x2010) = 0;
   *(ch_state + 0x2020) = 0;
   *(ch_state + 0x2028) = 0;
